requirements.txt
asgiref==3.4.1
Django==3.2.7
django-filter==21.1
psycopg2==2.7.7
pytz==2021.1
sqlparse==0.4.2
typing-extensions==3.10.0.2
